* SSE-ized ray/tri code. need to test some more., other changes
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@37 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/raymesh.cpp
@@ -111,7 +111,7 @@ void RayMesh::build(bool use_cached_trees)
 	//------------------------------------------------------------------------
 	conPrint("Building Mesh '" + name + "'...");
 	conPrint("\t" + toString(num_vertices) + " vertices (" + ::getNiceByteSize(vertex_data.size()*sizeof(float)) + ")");
-	conPrint("\t" + toString(triangles.size()) + " triangles (" + ::getNiceByteSize(triangles.size()*sizeof(SimpleIndexTri)) + ")");
+	conPrint("\t" + toString((unsigned int)triangles.size()) + " triangles (" + ::getNiceByteSize(triangles.size()*sizeof(SimpleIndexTri)) + ")");
 
 	if(RendererSettings::getInstance().cache_trees && use_cached_trees)
 	{
